% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anisotropy_index.R
\name{anisotropy_index}
\alias{anisotropy_index}
\title{Anisotropy index}
\usage{
anisotropy_index(probability_vector)
}
\arguments{
\item{probability_vector}{vector with length S whose elements are the
average probabilities of extinction of the species in a given community,
where S is the number of species.}
}
\value{
A non-negative number between 0 and 1 (isotropic community).
}
\description{
Given a vector whose elements are the average probabilities of extinction of
the species in a given community, this function calculates its anisotropy
index (i.e., the relative Shannon diversity index of such probabilities).
}
\examples{
probability_vector <- c(.25,.15,.60)
anisotropy_index(probability_vector)

}
