% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anisotropy_metrics_4_int_matrix.R
\name{anisotropy_metrics_4_int_matrix}
\alias{anisotropy_metrics_4_int_matrix}
\title{Anisotropy metrics for a given interaction matrix}
\usage{
anisotropy_metrics_4_int_matrix(
  A_int,
  number_Omega_replicates = 1000,
  number_boot_replicates = 1000,
  use_chol_decomp = F
)
}
\arguments{
\item{A_int}{a SxS interaction matrix, where S is the number of species.}

\item{number_Omega_replicates}{a number that specifies how many estimations
of Omega will be calculated by the quasi-Monte Carlo method in function
boot_prob_excl_Omega_raw. By default, this parameter is set to 1,000
replicates.}

\item{number_boot_replicates}{a number that specifies how many bootstrap
estimations of the probability of exclusion and Omega will be calculated
for each species in function boot_prob_excl_Omega_raw. By default, this
parameter is set to 1,000 replicates.}

\item{use_chol_decomp}{boolean that specifies if the QR decomposition should
be used to compute Omega.}
}
\value{
A tibble with 1 row and 6 columns, that contain the following data:
\itemize{
\item The mean value of small Omega.
\item The lower bound of the 95\% confidence interval for small Omega.
\item The upper bound of the 95\% confidence interval for small Omega.
\item The mean value of the anisotropy index.
\item The lower bound of the 95\% confidence interval for the anisotropy
index.
\item The upper bound of the 95\% confidence interval for the anisotropy
index.
}
}
\description{
Given an interaction matrix, this function uses the output of the function
boot_prob_excl_Omega_raw to estimate the mean values and 95\% confidence
intervals of the following metrics: small omega (i.e., the S root of capital
Omega, that is, the proportion of the feasible parameter space inside the
unit sphere) and the anisotropy index (i.e., the relative Shannon diversity
index of the species' probabilities of exclusion).
}
\examples{
numCores <- detectCores()
numCores
registerDoParallel(2)
A_int <- -1*diag(c(3,5,7))
anisotropy_metrics_4_int_matrix(A_int)
stopImplicitCluster()

}
\references{
\url{https://doi.org/10.1016/j.jtbi.2018.04.030}
}
