% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_omega.R
\name{calculate_omega}
\alias{calculate_omega}
\title{Calculate omega of arbitrary region}
\usage{
calculate_omega(vertex, raw = FALSE, nsamples = 1000, method = "convex_hull")
}
\arguments{
\item{vertex}{matrix of the vertices}

\item{raw}{logical, if TRUE, return the raw value of omega, default is FALSE}

\item{nsamples}{number of uniform samples when using "convex_hull" method, default is 1000}

\item{method}{method to calculate omega with "convex_hull" or "sphere", default is "convex_hull"}
}
\value{
the Omega value
}
\description{
function that computes the normalized feasibility from an interaction matrix
}
