% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_negative_definite.R
\name{check_negative_definite}
\alias{check_negative_definite}
\title{Check if a matrix is negative definite}
\usage{
check_negative_definite(matA)
}
\arguments{
\item{matA}{Numeric, an interaction matrix}
}
\value{
TRUE if the matrix is negative definite, FALSE otherwise
}
\description{
Check if a matrix is negative definite
}
\examples{
matA <- generate_inte_rand(4, 1, 1, 0, "norm")  ## Generate a random interaction matrix
check_negative_definite(matA)  ## Check if the matrix is negative definite
matB <- generate_inte_gs(4, 1, 1, 0, "norm")  ## Generate a globally stable random interaction matrix
check_negative_definite(matB)  ## Check if the matrix is negative definite
}
