% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_to_exclusion.R
\name{closest_to_exclusion}
\alias{closest_to_exclusion}
\title{Parameters of the species closest to exclusion}
\usage{
closest_to_exclusion(center, A_int)
}
\arguments{
\item{center}{an S vector that contains the initial intrinsic growth
rates that are used to computed the minimal isotropic variation that
excludes one or more species from the community.}

\item{A_int}{SxS interaction matrix, where S is the number of species.}
}
\value{
A matrix with 2xS columns. The number of rows can take values
between 1 and S. Each row contains the information of one of species that
is closest to exclusion. Specifically, it shows:
\itemize{
\item The intrinsic growth rate of the species when it is excluded
(r_T_i).
\item The abundance of species when it is excluded (N_T_i).
}
}
\description{
Given an interaction matrix and an initial vector of intrinsic growing rates,
denoted by center, this function computes the intrinsic growing
rates and abundances of the species that is/are closest to exclusion.
}
\examples{
A_int <- (-1) * diag(c(3,5,7))
incenter_inradius_isoprob <- incenter_inradius_isoprob_calculation(A_int)
I <- incenter_inradius_isoprob[[1]]
center <- I + runif(nrow(A_int))
closest_to_exclusion(center, A_int)

}
