% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone_vertices_director_vertices.R
\name{cone_vertices_director_vertices}
\alias{cone_vertices_director_vertices}
\title{Vertices of the feasibility domain}
\usage{
cone_vertices_director_vertices(A_int)
}
\arguments{
\item{A_int}{SxS interaction matrix, where S is the number of species.}
}
\value{
Dataframe with S rows and 2xS+2 columns. Each row contains the vertex
information for each species, when considering the intersection
of the feasibility domain and the unit ball. Specifically, each row
shows:
\itemize{
\item The intrinsic growth rate of species i at the given vertex (r_i).
\item The vertex ID.
\item The vertex name.
\item the normal vector of the great circle that contains all the S-1
remaining vertices and the origin of the unit ball.
}
}
\description{
This function computes the vertices of the simplicial cone of the
feasibility domain, when it intersects the unit ball, as well as the normal
vector of the great circle that contains all the S-1 remaining vertices and
the origin of the unit ball (i.e., a point where all the intrinsic growth
rates are equal to zero).
}
\examples{
A_int <- -1*diag(c(3,5,7))
cone_vertices_director_vertices(A_int)

}
