% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeasabilityR-package.R
\docType{package}
\name{feasabilityR-package}
\alias{feasabilityR}
\alias{feasabilityR-package}
\title{feasabilityR: Community ecology feasability analysis}
\description{
Analysis of feasibility within the largest framework of structural stability in ecology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MITEcology/feasibility_analysis}
  \item Report bugs at \url{https://github.com/MITEcology/feasibility_analysis/issues}
}

}
\author{
\strong{Maintainer}: Sergei Saavedra \email{nacho.bartomeus@gmail.com} (\href{https://orcid.org/0000-0000-0000-0000 }{ORCID})

Authors:
\itemize{
  \item Oscar Godoy \email{email@dot.com} (\href{https://orcid.org/0000-0000-0000-0000 }{ORCID})
  \item Ignasi Bartomeus \email{email@dot.com} (\href{https://orcid.org/0000-0000-0000-0000 }{ORCID})
  \item xxxxx xxxxxx \email{email@dot.com} (\href{https://orcid.org/0000-0000-0000-0000 }{ORCID})
}

}
\keyword{internal}
