% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility_asymmetry.R
\name{feasibility_asymmetry}
\alias{feasibility_asymmetry}
\title{Calculate the asymmetry of the feasibility domain}
\usage{
feasibility_asymmetry(matA)
}
\arguments{
\item{matA}{Numeric, an SxS interaction matrix A}
}
\value{
the asymmetry value
}
\description{
Calculate the asymmetry of the feasibility domain of community \eqn{\mathcal{S}}.
}
\note{
the asymmetry value is defined as standard deviation of \eqn{\{||v_1||, ||v_2||, \cdots, ||v_S||\}}, where \eqn{v_i} is the \eqn{i}-th column of the interaction matrix, i.e. the \eqn{i}-th vertex of the feasibility domain \eqn{D(\mathbf{S})}. Note that \eqn{v_i} is not normalized.
}
\examples{
matA <- generate_inte_rand(4, 1, 1, "norm")  ## Generate a random interaction matrix
feasibility_asymmetry(matA)
}
