% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility_community.R
\name{feasibility_community}
\alias{feasibility_community}
\title{Calculate the feasibility of a community S with n species}
\usage{
feasibility_community(matA, nt = 30, raw = TRUE)
}
\arguments{
\item{matA}{Numeric, an nxn interaction matrix A}

\item{nt}{Numeric, number of replications to reduce numerical instabilities, default to 30}

\item{raw}{Logical, whether to return the raw feasibility or the normalized feasibility, default to TRUE}
}
\value{
Feasibility of all the n species in community S.

Note Inside the main function raw and nt can be changed.
}
\description{
This estimates the feasibility domain generated by an interaction matrix of n species. This measure typically decreases with dimension n. If the matrix has positive and negative values then feasibility will be within \link{0,0.5}; otherwise \link{0,1/2^n}---these bounds are important if the user aims to transform feasibility into a probability measure that assumes a uniform distribution of directions in parameter space.
}
\examples{
matA <- generate_inte_rand(4, 1, 1, "norm")  ## Generate a random interaction matrix of 4 species
nA <- nrow(matA) ## total number of species in the interaction matrix A
feasibility_community(matA) ## This is the feasibility of the community A with 4 species. This measure cannot be compared across communities with different number of species.
(feasibility_community(matA)) ^ (1 / nA) ## This is the species-specific feasibility of the community A with 4 species. This measure can be conmpared across communities with different number of species.
}
