% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility_contribution.R
\name{feasibility_contribution}
\alias{feasibility_contribution}
\title{Calculate contribution to the feasibility of a community S by a species i}
\usage{
feasibility_contribution(matA, sp, nt = 30)
}
\arguments{
\item{matA}{Numeric, a nxn interaction matrix A}

\item{sp}{Numeric, index of the focal species}

\item{nt}{Numeric, number of replications to reduce numerical instabilities, default to 30}
}
\value{
Contribution of species i to the feasibility of a community
}
\description{
The ratio between the feasibility with n species and with n-1 species. How much the feasibility of a community S is positively or negatively affected by a species i. Outcomes above (resp. below) 1 mean a positive (resp. negative) contribution.
}
\examples{
matA <- generate_inte_gs(4, 1, 1, "norm")  ## Generate a random interaction matrix of nA=4 species
feasibility_contribution(matA, sp = 1) ## Contribution of species 1 to the feasibiltiy of the entire community of nA=4 species.
}
