% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility_invasion.R
\name{feasibility_invasion}
\alias{feasibility_invasion}
\title{Calculate feasibility of invasion to a community by a species i}
\usage{
feasibility_invasion(matA, inv, nt = 5000)
}
\arguments{
\item{matA}{Numeric, a nxn interaction matrix A}

\item{inv}{Numeric, index of the invading species}

\item{nt}{Numeric, number of sampling points, default to 5000}
}
\value{
Feasibility of invasion of species i to a community of n-1 species
}
\description{
Estimates the feasibility that a species i can invade a community of n-1 species. Feasibility of the domain of positive intrinsic growth rate (denoted by IGR_i>0). This considers that the rest of the species (\eqn{j \neq i \in \mathcal{S}}) are already at equilibrium. Recall that this can be transformed into a probability.
}
\examples{
matA <- generate_inte_gs(4, 1, 1, "norm")  ## Generate a random interaction matrix of nA=4 species
feasibility_invasion(matA, inv = 1) ## Estimates the feasibility of invasion by species 1 to the nA-1 species.
}
