% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility_overlap.R
\name{feasibility_overlap}
\alias{feasibility_overlap}
\title{Calculate the feasibility overlap between two communities}
\usage{
feasibility_overlap(matA, matB, raw = TRUE, nsamples = 3000, nt = 10)
}
\arguments{
\item{matA}{Numeric, an SxS interaction matrix A}

\item{matB}{Numeric, an SxS interaction matrix B}

\item{raw}{Logical, whether to return the raw feasibility (TRUE) or the normalized feasibility (FALSE), default to TRUE}

\item{nsamples}{Numeric, number of MonteCarlo sampling points, default to 3000}

\item{nt}{Numeric, number of replications to reduce numerical instabilities, default to TRUE}
}
\value{
A numeric value of the feasibility overlap between these two communities
}
\description{
Feasibility overlap between two communities of the same dimension.
}
\examples{
matA <- generate_inte_rand(4, 1, 1, "norm")  ## Generate a random interaction matrix A of dimension nA=4 species
matB <- generate_inte_rand(4, 1, 1, "norm")  ## Generate a random interaction matrix B of dimension nB=4 species
feasibility_overlap(matA,matB)
}
