% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility_partition.R
\name{feasibility_partition}
\alias{feasibility_partition}
\title{calculate the feasibility partition of all possible compositions}
\usage{
feasibility_partition(matA, nt = 30)
}
\arguments{
\item{matA}{Numeric, interaction matrix defining the dynamics of the system}

\item{nt}{Numeric, number of replications to reduce numerical instabilities, default to 30}
}
\value{
a list of all possible regions (as matrices)
}
\description{
calculate the feasibility partition of all possible compositions \eqn{\mathcal{C} \subseteq \mathcal{S}}
}
\note{
the interaction matrix matA needs to be globally stable in order for a meaningful partition of the parameter space.
}
\examples{
matA <- generate_inte_gs(4, 0.8, 1, "norm")  ## Generate a random interaction matrix
feasibility_partition(matA)
}
