% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility_resistance_partial.R
\name{feasibility_resistance_partial}
\alias{feasibility_resistance_partial}
\title{Calculate the partial resistance of a community to parameter perturbations}
\usage{
feasibility_resistance_partial(matA, r, norm = "l2")
}
\arguments{
\item{matA}{the interaction matrix}

\item{r}{the intrinsic growth rate}

\item{norm}{the norm to embed feasibility domain, either "l1" or "l2"}
}
\value{
a numeric value of the partial (some but not all species) resistance
}
\description{
partial resistance of community to parameter perturbations. Partial resistance is defined as the closest distance between the intrinsic growth rate vector to any of the vertices of the feasibility domain.
}
\examples{
matA <- generate_inte_rand(4, 1, 1, "norm") ## Generate random interaction matrix of nA=4 species
r <- matA \%*\% c(runif(4, 0, 1)) ## Generate random location of the community inside the feasibility domain
feasibility_resistance_partial(matA, r) ## The max resistance to parameter perturbations before losing nA-1 species.
}
\seealso{
feasibility_resistance_full
}
