% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility_species.R
\name{feasibility_species}
\alias{feasibility_species}
\title{Calculate the feasibility of individual species}
\usage{
feasibility_species(matA, sp, nt = 30)
}
\arguments{
\item{matA}{the interaction matrix of the community}

\item{sp}{the index of the focal species}
}
\value{
the feasibility value of species \eqn{sp}
}
\description{
calculate the feasibility of species \eqn{sp}'s survival inside a community \eqn{\mathcal{S}}
}
\note{
the feasibility of each individual species \eqn{sp} considering that no species in the community is already at equilibrium. This is the union of feasibility regions related to each species \eqn{sp}. For example, the feasibility of species \eqn{1} is the summation of the size of the feasibility regions \eqn{D(1,2,3)}, \eqn{D(1,2,\overline{3})} \eqn{D(1,\overline{2},3)} and \eqn{D(1,\overline{2},\overline{3})}. Note that \eqn{\overline{\cdot}} corresponds to the absence of a given species.
}
\examples{
matA <- generate_inte_gs(4, 0.8, 1, "norm")  ## Generate a random interaction matrix
feasibility_species(matA, 2)
}
