% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility_transition.R
\name{feasibility_transition}
\alias{feasibility_transition}
\title{Calculate the transition matrix for switching behavior given interaction matrix}
\usage{
feasibility_transition(inte, beta)
}
\arguments{
\item{inte}{Numeric, interaction matrix defining the dynamics of the system}

\item{beta}{Numeric, structual parameter for the transition matrix, default to 1}
}
\value{
a row-normalized transition matrix among compositions
}
\description{
calculate the feasibility partition of all possible 2^n compositions of a pool of n species assuming Lotka-Volterra dynamics. Then calculate the transition matrix based on the regions vector and distances matrix with specified structual parameters.
}
\examples{
matA <- generate_inte_gs(3, 1, 1, "norm")  ## Generate a random interaction matrix with nA=5 species
feasibility_transition(matA, 1) ## Transition matrix among the 2^3=8 possible compositions of the 3-taxa system

}
