% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeasibilityR-package.R
\docType{package}
\name{feasibilityR-package}
\alias{feasibilityR}
\alias{feasibilityR-package}
\title{feasibilityR: Community ecology feasibility analysis}
\description{
Analysis of feasibility within the framework of structural stability in ecology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MITEcology/feasibility_analysis}
  \item Report bugs at \url{https://github.com/MITEcology/feasibility_analysis/issues}
}

}
\author{
\strong{Maintainer}: Serguei Saavedra \email{sersaa@mit.com} (\href{https://orcid.org/0000-0003-1768-363X }{ORCID})

Authors:
\itemize{
  \item Oscar Godoy \email{ogodoy.re@gmail.com} (\href{https://orcid.org/0000-0003-4988-6626 }{ORCID})
  \item Ignasi Bartomeus \email{nacho.bartomeus@gmail.com} (\href{https://orcid.org/0000-0001-7893-4389 }{ORCID})
  \item Jie Deng \email{jiedeng@mit.edu} (\href{https://orcid.org/0000-0002-0604-7884 }{ORCID})
  \item Chengyi Long \email{cylong@mit.edu} (\href{https://orcid.org/0000-0002-6005-3740 }{ORCID})
}

Other contributors:
\itemize{
  \item Chuliang Song \email{clsong.ecology@gmail.com} (\href{https://orcid.org/0000-0001-7490-8626 }{ORCID}) [contributor]
}

}
\keyword{internal}
