% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_inte_ev.R
\name{generate_inte_ev}
\alias{generate_inte_ev}
\title{Generate random symmetric interaction matrix with speficied eigenvalues}
\usage{
generate_inte_ev(S, lambda, lmean = -0.5, loffset = 0.5, rep = 1)
}
\arguments{
\item{S}{Numeric, number of species}

\item{lmean}{Numeric, center of the eigenvalues range}

\item{loffset}{Numeric, half length of the eigenvalues range}

\item{rep}{Numeric, number of matrices wanted, default to 1}

\item{lamba}{Numberic, either specifing the vector of eigenvalues, or the method to sample eigenvalues}
}
\value{
a single matrix or a list of SxS interaction matrices A.
}
\description{
Generate random symmetric interaction matrix with speficied eigenvalues
}
\note{
This approach decompose a symmetric interaction matrix to a product of random orthogonal matrix and specified eignevalues. Thus, it can be used to explictly assign properties or the exact eigenvalues of the interaction matrix. However, the interaction matrix has to be symmetric.
}
\examples{
matA <- generate_inte_ev(5, c(-1, -2, -3, -4, -5)) #assigning eigenvalues
check_negative_definite(matA)

matA <- generate_inte_ev(4, "unif", 0.5, 0.5) #uniformly sampling eigenvalues
check_negative_definite(matA)

matA <- generate_inte_ev(4, "lognorm", 0.5, 0.5) #lognormal sampling eigenvalues
check_negative_definite(matA)

}
