% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_inte_gs.R
\name{generate_inte_gs}
\alias{generate_inte_gs}
\title{Generate random interaction matrix that satisfies global dynamical stability}
\usage{
generate_inte_gs(
  S,
  sigma,
  conne = 1,
  dist = "norm",
  mu = -0.1 * sigma,
  trail = 0
)
}
\arguments{
\item{S}{Numeric, number of species}

\item{sigma}{Numeric, standard deviation of interaction}

\item{conne}{Numeric, connectivity of interaction, default to 1}

\item{dist}{Character, distribution of interaction, default to "norm"}

\item{mu}{Numeric, mean interaction, default to 0}

\item{trail}{Numeric, starting number of trails to generate negative definite matrix, default to 0}
}
\value{
a globally stable SxS interaction matrix A.
}
\description{
Generate a globally stable random interaction matrix A of a community S
}
\note{
Note that if one chooses a half-normal distribution, the interaction matrix can be interpreted as a purely competition system. As default, the diagonal values are all set to a_ii = -1 (e.g., self-regulation).
}
\examples{
matA <- generate_inte_gs(4, 1, 1, "norm")  ## Generate a globally stable random interaction matrix
matA
}
