% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compo.R
\name{get_compo}
\alias{get_compo}
\title{Get compositions of a system}
\usage{
get_compo(num, nv = 0)
}
\arguments{
\item{num}{Numeric, number of species in the system}

\item{nv}{Numeric, value to fill in the empty table and represents absence, default to 0}
}
\value{
the table of all possible communities (as a matrix)
}
\description{
function that generates a table of presence/absence combinations
}
