% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effective_competition.R
\name{get_effective_competition}
\alias{get_effective_competition}
\title{Get effective competition matrix}
\usage{
get_effective_competition(matA)
}
\arguments{
\item{matA}{matrix of bipartite network (?)}
}
\value{
effective competition matrix
}
\description{
TBA
}
\examples{
# TBA

}
