% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incenter_inradius_isoprob_calculation.R
\name{incenter_inradius_isoprob_calculation}
\alias{incenter_inradius_isoprob_calculation}
\title{Main parameters of the maximum isotropic cap}
\usage{
incenter_inradius_isoprob_calculation(A_int)
}
\arguments{
\item{A_int}{a SxS interaction matrix, where S is the number of species.}
}
\value{
A list with 3 elements:
\itemize{
\item The first element is a vector that contains the intrinsic growth
rates of the feasibility domain's incenter, I.
\item The second element is a number that represents the great circle
distance from I to the border of the maximum  isotropic cap, theta.
\item The third element is a number that shows area of the maximum
isotropic cap, Xi.
}
}
\description{
Given an interaction matrix, this function computes the main parameters of
its maximum isotropic cap.
}
\examples{
A_int <- -1*diag(c(3,5,7))
incenter_inradius_isoprob_calculation(A_int)

}
