% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_detection.R
\name{inside_detection}
\alias{inside_detection}
\title{function that detects if a point is inside a simplex}
\usage{
inside_detection(Span, vector)
}
\arguments{
\item{Span}{matrix of the simplex}

\item{vector}{vector to be tested}
}
\value{
1 if the point is inside the simplex, 0 otherwise
}
\description{
This function detects if a point is inside a simplex
}
