% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_vertex_detection.R
\name{inside_vertex_detection}
\alias{inside_vertex_detection}
\title{Detect inside vertex}
\usage{
inside_vertex_detection(matA, matB)
}
\arguments{
\item{matA}{matrix of the original vertexes}

\item{matB}{matrix of the extreme points}
}
\value{
a list with the extreme points that belong to original vertexes
}
\description{
function that computes all the extreme points that belong to original vertexes
}
