% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection_vertex_detection.R
\name{intersection_vertex_detection}
\alias{intersection_vertex_detection}
\title{Detect intersection vertex}
\usage{
intersection_vertex_detection(S, M)
}
\arguments{
\item{S}{matrix of the first set of vertexes}

\item{M}{matrix of the second set of vertexes}
}
\value{
a list with the extreme points that intersects these cones
}
\description{
function that computes all the extreme points generated that are generated by the intersections of the cones
}
