% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_min_intrinsic_growth_rates.R
\name{max_min_intrinsic_growth_rates}
\alias{max_min_intrinsic_growth_rates}
\title{Maximum and minimum (feasible) intrinsic growth rates for the community's
species}
\usage{
max_min_intrinsic_growth_rates(A_int, number_points_per_species = 1000)
}
\arguments{
\item{A_int}{SxS interaction matrix, where S is the number of species.}

\item{number_points_per_species}{number of intrinsic growth rates vectors to
be randomly sampled in the border of the feasibility domain where a given
species goes excluded.}
}
\value{
Dataframe with S rows and 3 columns. Each row contains the
information for each species. Specifically, each row shows:
\itemize{
\item The species name.
\item The minimum intrinsic growth rate estimated for such species.
\item The miaximum intrinsic growth rate estimated for such species.
}
}
\description{
This function computes a rough estimation of the species range of feasible
intrinsic growth rates in a community with S species, under Lotka-Volterra
dynamics. To do so, for each species, we randomly select 1,000 intrinsic
growth rates vectors in the border where a given species goes extinct, and
then we extract the the minimum and maximum values for each component of the
1,000xS selected vectors.
}
\examples{
A_int <- -1*diag(c(3,5,7))
max_min_intrinsic_growth_rates(A_int, number_points_per_species = 1000)

}
