% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm1.R
\name{norm1}
\alias{norm1}
\title{function that normalizes a vector in the L1 norm}
\usage{
norm1(a)
}
\arguments{
\item{a}{Numeric, a vector or a matrix}
}
\value{
A vector or a matrix, normalized in the L1 norm
}
\description{
This function normalizes a vector in the L1 norm.
}
