% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm2.R
\name{norm2}
\alias{norm2}
\title{Normalizes vector(s) in the L2 norm}
\usage{
norm2(a)
}
\arguments{
\item{a}{Numeric, a vector or a matrix}
}
\value{
A vector or a matrix, normalized in the L2 norm
}
\description{
This function normalizes vector(s) in the L2 norm.
}
