% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_extinction_4_int_matrix.R
\name{prob_extinction_4_int_matrix}
\alias{prob_extinction_4_int_matrix}
\title{Probabilities of exclusion}
\usage{
prob_extinction_4_int_matrix(
  A_int,
  number_Omega_replicates = 1000,
  number_boot_replicates = 1000,
  use_chol_decomp = F
)
}
\arguments{
\item{A_int}{a SxS interaction matrix, where S is the number of species.}

\item{number_Omega_replicates}{a number that specifies how many estimations
of Omega will be calculated by the quasi-Monte Carlo method in function
boot_prob_excl_Omega_raw. By default, this parameter is set to 1,000
replicates.}

\item{number_boot_replicates}{a number that specifies how many bootstrap
estimations of the probability of exclusion and Omega will be calculated
for each species in function boot_prob_excl_Omega_raw. By default, this
parameter is set to 1,000 replicates.}

\item{use_chol_decomp}{boolean that specifies if the QR decomposition should
be used to compute Omega.}
}
\value{
A tibble with S rows and 4 columns, that contain the following data:
\itemize{
\item The species ID.
\item The mean value of the species' probability of exclusion.
\item The lower bound of the 95\% confidence interval for the species'
probability of exclusion.
\item The upper bound of the 95\% confidence interval for the species'
probability of exclusion.
}
}
\description{
Given an interaction matrix, this function uses the output of the function
boot_prob_excl_Omega_raw to estimate the mean values and 95\% confidence
intervals of the species' probabilities of exclusion.
}
\examples{
numCores <- detectCores()
numCores
registerDoParallel(2)
A_int <- -1*diag(c(3,5,7))
prob_extinction_4_int_matrix(A_int)
stopImplicitCluster()

}
\references{
\url{https://doi.org/10.1016/j.jtbi.2018.04.030}
}
