% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runif_on_sphere.R
\name{runif_on_sphere}
\alias{runif_on_sphere}
\title{Uniform random points on a sphere}
\usage{
runif_on_sphere(n, d, r = 1)
}
\arguments{
\item{n}{Integer, the number of points to generate}

\item{d}{Integer, the dimension of the sphere}

\item{r}{Numeric, the radius of the sphere}
}
\value{
A matrix of size n * d, each row is a point on the sphere
}
\description{
Generate uniform random points on a sphere
}
\examples{
runif_on_sphere(20, 4, 1)
}
