% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_distance_to_exclusion.R
\name{sp_distance_to_exclusion}
\alias{sp_distance_to_exclusion}
\title{Parameters of the species closest to exclusion}
\usage{
sp_distance_to_exclusion(center, A_int)
}
\arguments{
\item{center}{an S vector that contains the initial intrinsic growth
rates that are used to computed the minimal isotropic variation that
excludes one or more species from the community.}

\item{A_int}{SxS interaction matrix, where S is the number of species.}
}
\value{
A tibble data frame with S rows and 2 columns. Each row contains the
species name and its great circle distance to the border where it goes
excluded (i.e, that distance is the absolute value of the angle between
the center and the tangent point in a given "edge" of the feasibility domain).
}
\description{
Given an interaction matrix and an initial vector of intrinsic growing rates,
denoted by center, this function computes the arc distance from each species
to the feasibility domain's border where only such species is excluded.
}
\examples{
A_int <- (-1) * diag(c(3,5,7))
incenter_inradius_isoprob <- incenter_inradius_isoprob_calculation(A_int)
I <- incenter_inradius_isoprob[[1]]
center <- I + runif(nrow(A_int))
sp_distance_to_exclusion(center, A_int)

}
