% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex_detection.R
\name{vertex_detection}
\alias{vertex_detection}
\title{detect all extreme points}
\usage{
vertex_detection(A, B)
}
\arguments{
\item{A}{matrix of the first set of vertexes}

\item{B}{matrix of the second set of vertexes}
}
\value{
a matrix of all the extreme points
}
\description{
function that computes all the extreme points
}
