% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertices_unit_ball.R
\name{vertices_unit_ball}
\alias{vertices_unit_ball}
\title{Vertices of the feasibility domain}
\usage{
vertices_unit_ball(A_int)
}
\arguments{
\item{A_int}{SxS interaction matrix, where S is the number of species.}
}
\value{
Dataframe with S rows and 2xS+1 columns. Each row contains the vertex
information for each species, when considering the intersection
of the feasibility domain and the unit ball. Specifically, each row
shows:
\itemize{
\item The abundance of species i at the given vertex (N_i).
\item The intrinsic growth rate of species i at the given vertex (r_i).
\item The vertex ID.
}
}
\description{
This function computes the vertices of the simplicial cone of the
feasibility domain when it intersects the unit ball.
}
\examples{
A_int <- -1*diag(c(3,5,7))
vertices_unit_ball(A_int)

}
