% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility_asymmetry.R
\name{feasibility_asymmetry}
\alias{feasibility_asymmetry}
\title{Calculate the asymmetry of the feasibility domain of community S with n species}
\usage{
feasibility_asymmetry(matA)
}
\arguments{
\item{matA}{Numeric, an SxS interaction matrix A}
}
\value{
the asymmetry value
}
\description{
The asymmetry value is defined as the standard deviation of \eqn{\{||v_1||, ||v_2||, \cdots, ||v_n||\}}, where \eqn{v_i} is the \eqn{i}-th column of the interaction matrix, i.e. the \eqn{i}-th vertex of the feasibility domain. Note that \eqn{v_i} is not normalized.
think about a triangle as the feasibility domain. The asymmetry measures the variation in the legth of borders. Each of the borders represent the feasbility of the n-1 communities.
}
\examples{
matA <- generate_inte_rand(4, 1, 1, "norm") ## Generate a random matrix of 4 species following a normal distribution with mean=1 and sd=1
feasibility_asymmetry(matA) ## This measures the irregularity of the feasiblity domain of the community A with 4 species. The larger the outcome, the larger the assymetry
}
