% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility_contribution.R
\name{feasibility_contribution}
\alias{feasibility_contribution}
\title{Calculate contribution to the feasibility of a community by a species}
\usage{
feasibility_contribution(matA, sp, nt = 30)
}
\arguments{
\item{matA}{Numeric, an \eqn{S\times S} interaction matrix \eqn{A}}

\item{sp}{Numeric, index of the focal species}

\item{nt}{Numeric, number of replications to reduce numerical instabilities, default to 30}
}
\value{
How much the feasibility of a community S is positively or negatively affected by a species i. Outcomes above (resp. below) 1 mean a positive (resp. negative) contribution.
}
\description{
The ratio between the feasibility with \eqn{S} species and with \eqn{S-1} species.
}
\examples{
matA <- generate_inte_gs(4, 1, 1, "norm")  ## Generate a random interaction matrix
feasibility_contribution(matA, sp = 1)
}
