% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility_invasion.R
\name{feasibility_invasion}
\alias{feasibility_invasion}
\title{Calculate feasibility of invasion to a community by a species}
\usage{
feasibility_invasion(matA, inv, nt = 5000)
}
\arguments{
\item{matA}{Numeric, an SxS interaction matrix A}

\item{inv}{Numeric, index of the invading species}

\item{nt}{Numeric, number of sampling points, default to 5000}
}
\value{
Feasibility of the domain of positive intrinsic growth rate (denoted by \eqn{IGR_i>0}). This considers that the rest of the species (\eqn{j \neq i \in \mathcal{S}}) are already at equilibrium.
}
\description{
Calculate feasibility of invasion to a community by a species
}
\examples{
matA <- generate_inte_gs(4, 1, 1, "norm")  ## Generate a random interaction matrix
feasibility_invasion(matA, inv = 1)
}
