% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility_overlap.R
\name{feasibility_overlap}
\alias{feasibility_overlap}
\title{Feasibility overlap between communities \eqn{\mathcal{S}_1} and \eqn{\mathcal{S}_2}}
\usage{
feasibility_overlap(matA, matB, raw = TRUE, nsamples = 3000, nt = 10)
}
\arguments{
\item{matA}{Numeric, an SxS interaction matrix A}

\item{matB}{Numeric, an SxS interaction matrix B}

\item{raw}{Logical, whether to return the raw feasibility (TRUE) or the normalized feasibility (FALSE), default to TRUE}

\item{nsamples}{Numeric, number of MonteCarlo sampling points, default to 3000}

\item{nt}{Numeric, number of replications to reduce numerical instabilities, default to TRUE}
}
\value{
Overlap: feasiblity region of overlap
}
\description{
Feasibility overlap between communities \eqn{\mathcal{S}_1} and \eqn{\mathcal{S}_2}
}
\note{
interaction matrices need to have same dimension.

Note Inside the function raw and nt can be changed.
}
\examples{
matA <- generate_inte_rand(4, 1, 1, "norm")  ## Generate a random interaction matrix A
matB <- generate_inte_rand(4, 1, 1, "norm")  ## Generate a random interaction matrix B
feasibility_overlap(matA,matB)
}
