% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility_partition.R
\name{feasibility_partition}
\alias{feasibility_partition}
\title{Calculate the feasibility partition of all possible compositions}
\usage{
feasibility_partition(matA, nt = 30)
}
\arguments{
\item{matA}{Numeric, interaction matrix defining the dynamics of the system}

\item{nt}{Numeric, number of replications to reduce numerical instabilities, default to 30}
}
\value{
a vector of length 2^n for all possible compositions.
}
\description{
calculate the feasibility partition of all possible 2^n compositions of a pool of n species. The interaction matrix matA needs to be globally stable in order for a meaningful partition of the parameter space. The sum should be close to one.
}
\examples{
matA <- generate_inte_gs(4, 1, 1, "norm")  ## Generate a random interaction matrix with nA=4 species
feasibility_partition(matA) ## Feasibility of each of the 16 possible partitions
}
