% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibility_resistance_full.R
\name{feasibility_resistance_full}
\alias{feasibility_resistance_full}
\title{Calculate the full resistance to parameter perturbations}
\usage{
feasibility_resistance_full(matA, r, norm = "l2", nt = 100, all = FALSE)
}
\arguments{
\item{matA}{the interaction matrix}

\item{r}{the intrinsic growth rate}

\item{norm}{the norm to embed feasibility domain, either "l1" or "l2"}

\item{nt}{the number of sampling points on the border of feasibility domain}

\item{all}{whether to return all resistance values from the samples or taking the minimal value as the estimated resistance}
}
\value{
a numeric value of the full resistance
}
\description{
full resistance of community $\mathcal{S}$ to parameter perturbations
}
\note{
full resistance is the distance between an intrinsic growth rate vector inside the feasibility domain and a border of \eqn{D(\mathcal{S})}. It can be calculated by the nearest arc length (i.e. L2 norm) from \eqn{r} to a border as: \eqn{d_b = \arccos <r(N^*), r(\mathrm{border})>}; partial resistance is similarly defined as distance between the intrinsic growth rate vector to the vertices of feasibility domain.
}
\examples{
matA <- generate_inte_rand(4, 1, 1, "norm")
r <- matA \%*\% c(runif(4, 0, 1))
feasibility_resistance_full(matA, r)
}
\seealso{
\code{\link{feasibility_resistance_partial}}
}
